+++ src/De/de.controller.js
@@ -1,0 +1,4 @@
+function lancerDe(max){
+    return Math.floor((Math.random()*max)+1);
+}
+module.exports = lancerDe;
